Refactor Tabs.js to use hooks and split Tabs stories (#3463)
* Split Tabs stories


* Refactor Tab.js to use hooks


* Update type in type definitions


* defaultProps as default params & remove unused prop


* Merge stories for better readability


* Set defaultProps
