+++ src/js/components/Tabs/Tabs.js
@@ -1,4 +1,4 @@
-import React, { cloneElement, Children, Component } from 'react';
+import React, { cloneElement, Children, useState } from 'react';
 import { compose } from 'recompose';
 
 import { withTheme } from 'styled-components';
@@ -9,110 +9,92 @@ import { Box } from '../Box';
 
 import { StyledTabPanel, StyledTabs, StyledTabsHeader } from './StyledTabs';
 
-class Tabs extends Component {
-  static defaultProps = {
-    justify: 'center',
-    messages: {
-      tabContents: 'Tab Contents',
-    },
-    responsive: true,
-  };
+const Tabs = ({
+  children,
+  flex,
+  justify = 'center',
+  messages = { tabContents: 'Tab Contents' },
+  theme,
+  ...rest
+}) => {
+  const { activeIndex: propsActiveIndex, onActive } = rest;
+  const [activeIndex, setActiveIndex] = useState(rest.activeIndex || 0);
 
-  static getDerivedStateFromProps(nextProps, prevState) {
-    const { activeIndex } = nextProps;
-    const { activeIndex: stateActiveIndex } = prevState;
-    if (stateActiveIndex !== activeIndex && activeIndex !== undefined) {
-      return { activeIndex };
-    }
-    return { activeIndex: stateActiveIndex || 0 };
+  if (activeIndex !== propsActiveIndex && propsActiveIndex !== undefined) {
+    setActiveIndex(propsActiveIndex);
   }
 
-  state = {};
-
-  activateTab = index => {
-    const { activeIndex, onActive } = this.props;
-    if (activeIndex === undefined) {
-      this.setState({ activeIndex: index });
+  const activateTab = index => {
+    if (propsActiveIndex === undefined) {
+      setActiveIndex(index);
     }
     if (onActive) {
       onActive(index);
     }
   };
 
-  render() {
-    const {
-      children,
-      flex,
-      justify,
-      messages: { tabContents },
-      theme,
-      ...rest
-    } = this.props;
-    delete rest.activeIndex;
-    delete rest.onActive;
-    const { activeIndex } = this.state;
+  /* eslint-disable no-param-reassign */
+  delete rest.activeIndex;
+  delete rest.onActive;
+  /* eslint-enable no-param-reassign */
 
-    let activeContent;
-    let activeTitle;
-    const tabs = Children.map(
-      children,
-      (tab, index) => {
-        if (!tab) return undefined;
+  let activeContent;
+  let activeTitle;
+  const tabs = Children.map(
+    children,
+    (tab, index) => {
+      if (!tab) return undefined;
 
-        const tabProps = tab.props || {};
+      const tabProps = tab.props || {};
 
-        const isTabActive = index === activeIndex;
+      const isTabActive = index === activeIndex;
 
-        if (isTabActive) {
-          activeContent = tabProps.children;
-          if (typeof tabProps.title === 'string') {
-            activeTitle = tabProps.title;
-          } else {
-            activeTitle = index + 1;
-          }
+      if (isTabActive) {
+        activeContent = tabProps.children;
+        if (typeof tabProps.title === 'string') {
+          activeTitle = tabProps.title;
+        } else {
+          activeTitle = index + 1;
         }
+      }
 
-        return cloneElement(tab, {
-          active: isTabActive,
-          onActivate: () => this.activateTab(index),
-        });
-      },
-      this,
-    );
+      return cloneElement(tab, {
+        active: isTabActive,
+        onActivate: () => activateTab(index),
+      });
+    },
+    this,
+  );
 
-    const tabContentTitle = `${activeTitle || ''} ${tabContents}`;
+  const tabContentTitle = `${activeTitle || ''} ${messages.tabContents}`;
 
-    return (
-      <StyledTabs
+  return (
+    <StyledTabs
+      as={Box}
+      role="tablist"
+      flex={flex}
+      {...rest}
+      background={theme.tabs.background}
+    >
+      <StyledTabsHeader
         as={Box}
-        role="tablist"
-        flex={flex}
-        {...rest}
-        background={theme.tabs.background}
+        direction="row"
+        justify={justify}
+        flex={false}
+        wrap
+        background={theme.tabs.header.background}
+        gap={theme.tabs.gap}
       >
-        <StyledTabsHeader
-          as={Box}
-          direction="row"
-          justify={justify}
-          flex={false}
-          wrap
-          background={theme.tabs.header.background}
-          gap={theme.tabs.gap}
-        >
-          {tabs}
-        </StyledTabsHeader>
-        <StyledTabPanel
-          flex={flex}
-          aria-label={tabContentTitle}
-          role="tabpanel"
-        >
-          {activeContent}
-        </StyledTabPanel>
-      </StyledTabs>
-    );
-  }
-}
+        {tabs}
+      </StyledTabsHeader>
+      <StyledTabPanel flex={flex} aria-label={tabContentTitle} role="tabpanel">
+        {activeContent}
+      </StyledTabPanel>
+    </StyledTabs>
+  );
+};
 
+Tabs.defaultProps = {};
 Object.setPrototypeOf(Tabs.defaultProps, defaultProps);
 
 let TabsDoc;
